Remove 1.7 support
