--- src/main/java/fr/xephi/authme/listener/PlayerListener16.java
@@ -1,25 +1,0 @@
-package fr.xephi.authme.listener;
-
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerEditBookEvent;
-
-import javax.inject.Inject;
-
-/**
- * Listener of player events for events introduced in Minecraft 1.6.
- */
-public class PlayerListener16 implements Listener {
-
-    @Inject
-    private ListenerService listenerService;
-
-    @EventHandler(ignoreCancelled = true, priority = EventPriority.LOWEST)
-    public void onPlayerEditBook(PlayerEditBookEvent event) {
-        if (listenerService.shouldCancelEvent(event)) {
-            event.setCancelled(true);
-        }
-    }
-
-}
