--- src/main/java/fr/xephi/authme/listener/PlayerListener18.java
@@ -1,25 +1,0 @@
-package fr.xephi.authme.listener;
-
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerInteractAtEntityEvent;
-
-import javax.inject.Inject;
-
-/**
- * Listener of player events for events introduced in Minecraft 1.8.
- */
-public class PlayerListener18 implements Listener {
-
-    @Inject
-    private ListenerService listenerService;
-
-    @EventHandler(ignoreCancelled = true, priority = EventPriority.LOWEST)
-    public void onPlayerInteractAtEntity(PlayerInteractAtEntityEvent event) {
-        if (listenerService.shouldCancelEvent(event)) {
-            event.setCancelled(true);
-        }
-    }
-
-}
