+++ src/main/java/fr/xephi/authme/permission/PermissionsManager.java
@@ -455,11 +455,7 @@ public class PermissionsManager implements Reloadable {
      */
     public boolean loadUserData(OfflinePlayer offlinePlayer) {
         try {
-            try {
-                loadUserData(offlinePlayer.getUniqueId());
-            } catch (NoSuchMethodError e) {
-                loadUserData(offlinePlayer.getName());
-            }
+            loadUserData(offlinePlayer.getUniqueId());
         } catch (PermissionLoadUserException e) {
             ConsoleLogger.logException("Unable to load the permission data of user " + offlinePlayer.getName(), e);
             return false;
@@ -478,18 +474,5 @@ public class PermissionsManager implements Reloadable {
             return;
         }
         handler.loadUserData(uuid);
-    }
-
-    /**
-     * Loads the permission data of the given player name.
-     *
-     * @param name the name of the player.
-     * @throws PermissionLoadUserException if the action failed.
-     */
-    public void loadUserData(String name) throws PermissionLoadUserException {
-        if (!isEnabled()) {
-            return;
-        }
-        handler.loadUserData(name);
     }
 }
