+++ src/main/java/fr/xephi/authme/permission/handlers/LuckPermsHandler.java
@@ -196,14 +196,4 @@ public class LuckPermsHandler implements PermissionHandler {
             throw new PermissionLoadUserException("Unable to load the permission data of the user " + uuid, e);
         }
     }
-
-    @Override
-    public void loadUserData(String name) throws PermissionLoadUserException {
-        try {
-            UUID uuid = luckPermsApi.getUserManager().lookupUuid(name).get(5, TimeUnit.SECONDS);
-            loadUserData(uuid);
-        } catch (InterruptedException | ExecutionException | TimeoutException e) {
-            throw new PermissionLoadUserException("Unable to load the permission data of the user " + name, e);
-        }
-    }
 }
