+++ src/main/java/fr/xephi/authme/permission/handlers/PermissionHandler.java
@@ -109,7 +109,4 @@ public interface PermissionHandler {
 
     default void loadUserData(UUID uuid) throws PermissionLoadUserException {
     }
-
-    default void loadUserData(String name) throws PermissionLoadUserException {
-    }
 }
