+++ src/main/java/fr/xephi/authme/util/PlayerUtils.java
@@ -13,23 +13,6 @@ public final class PlayerUtils {
     }
 
     /**
-     * Get player's UUID if can, name otherwise.
-     *
-     * @param player Player to retrieve
-     *
-     * @return player's UUID or Name in String.
-     */
-    public static String getUuidOrName(OfflinePlayer player) {
-        // We may made this configurable in future
-        // so we can have uuid support.
-        try {
-            return player.getUniqueId().toString();
-        } catch (NoSuchMethodError ignore) {
-            return player.getName();
-        }
-    }
-
-    /**
      * Returns the IP of the given player.
      *
      * @param p The player to return the IP address for
